resolved PR comments
