+++ web-platform/website/src/components/pages/privacy-policy/PrivacyPolicyPageContentV2.tsx
@@ -1,4 +1,5 @@
 import { UserDocument } from "@mozilla/rally-shared-types/dist";
+import { useState } from "react";
 import { Container, Modal, ModalHeader } from "reactstrap";
 import { Button, Col, Row } from "reactstrap";
 import { style } from "typestyle";
@@ -22,15 +23,15 @@ import { PrivacyPolicySharing } from "./PrivacyPolicySharing";
 const strings = Strings.components.pages.privacyPolicy;
 const btnStrings = Strings.components.pages.privacyPolicy.buttons;
 
-export function PrivacyPolicyPageContentV2(props: { closeModal: object }) {
+export function PrivacyPolicyPageContentV2() {
   const { updateUserDocument, userDocument } = useUserDocument();
+  const [isOpen, setIsOPen] = useState(true);
 
   return (
     <Modal
       className={styles.modal}
       contentClassName={styles.modalContent}
-      isOpen={true}
-      {...props}
+      isOpen={isOpen}
     >
       <ModalHeader className={Fonts.Title}>{strings.modalHeader}</ModalHeader>
       <Container
@@ -57,7 +58,7 @@ export function PrivacyPolicyPageContentV2(props: { closeModal: object }) {
                 enrolled: true,
               });
 
-              props && props.closeModal && (props.closeModal as () => void)();
+              setIsOPen(false);
             }}
           >
             {btnStrings.v2.agree}
@@ -66,7 +67,7 @@ export function PrivacyPolicyPageContentV2(props: { closeModal: object }) {
             className={`d-flex fw-bold ps-4 pe-4 pt-2 pb-2 ${TertiaryButton} ${styles.button}`}
             outline
             onClick={() => {
-              props && props.closeModal && (props.closeModal as () => void)();
+              setIsOPen(false);
             }}
           >
             {btnStrings.v2.back}
