+++ web-platform/website/src/components/pages/studies/ProductToasts.tsx
@@ -1,31 +1,29 @@
 import { useEffect, useState } from "react";
-import { Alert, Button, Container } from "reactstrap";
+import { Container } from "reactstrap";
 import { style } from "typestyle";
 
 import { Strings } from "../../../resources/Strings";
 import { useAuthentication } from "../../../services/AuthenticationService";
 import { useStudies } from "../../../services/StudiesService";
 import { useUserDocument } from "../../../services/UserDocumentService";
-import { Colors, Spacing } from "../../../styles";
-import { ProductButton } from "../../../styles/Buttons";
-import { LinkStyles } from "../../../styles/LinkStyles";
-import { detectBrowser } from "../../../utils/BrowserDetector";
-import { BrowserType } from "../../../utils/BrowserType";
+import { Spacing } from "../../../styles";
+import { PrivacyPolicyPageContentV2 } from "../privacy-policy/PrivacyPolicyPageContentV2";
+import { ProductCheckEmailDialog } from "./ProductCheckEmailDialog";
+import { ToastComponent } from "./study-card/ToastComponent";
 
 const emailStrings = Strings.components.pages.studies.alerts.verifyEmail;
 const privacyStrings = Strings.components.pages.studies.alerts.privacyPolicy;
 const extensionStrings = Strings.components.pages.studies.alerts.addExtension;
 
-export function ProductToasts(props: {
-  openModal: object;
-  openPrivacyModal: object;
-}) {
+export function ProductToasts() {
   const [showEmailNotVerifiedToast, setShowEmailNotVerifiedToast] =
     useState<boolean>(false);
   const [showAddExtensionToast, setShowAddExtenionToast] =
     useState<boolean>(false);
   const [showAddPrivacyToast, setShowAddPrivacyToast] =
     useState<boolean>(false);
+  const [showEmailDialog, setShowEmailDialog] = useState<boolean>(false);
+  const [showPrivacyDialog, setShowPrivacyDialog] = useState<boolean>(false);
 
   const { isUserVerified, user } = useAuthentication();
   const { installedStudyIds } = useStudies();
@@ -36,9 +34,13 @@ export function ProductToasts(props: {
       setShowEmailNotVerifiedToast(true);
     }
 
-    if (!installedStudyIds.length) setShowAddExtenionToast(true);
+    if (!installedStudyIds.length) {
+      setShowAddExtenionToast(true);
+    }
 
-    if (userDocument && !userDocument.enrolled) setShowAddPrivacyToast(true);
+    if (userDocument && !userDocument.enrolled) {
+      setShowAddPrivacyToast(true);
+    }
   }, [isUserVerified, user, installedStudyIds]);
 
   return (
@@ -46,7 +48,7 @@ export function ProductToasts(props: {
       {showAddPrivacyToast && (
         <ToastComponent
           {...privacyStrings}
-          openPrivacyModal={props.openPrivacyModal}
+          openPrivacyModal={() => setShowPrivacyDialog(true)}
         />
       )}
 
@@ -55,97 +57,15 @@ export function ProductToasts(props: {
       )}
 
       {showEmailNotVerifiedToast && (
-        <ToastComponent {...emailStrings} openModal={props.openModal} />
+        <ToastComponent
+          {...emailStrings}
+          openModal={() => setShowEmailDialog(true)}
+        />
       )}
+
+      {showEmailDialog && <ProductCheckEmailDialog />}
+      {showPrivacyDialog && <PrivacyPolicyPageContentV2 />}
     </Container>
-  );
-}
-
-interface ToastProps {
-  icon: string;
-  text: string;
-  button: string;
-  type: string;
-  close?: string;
-  openModal?: object;
-  openPrivacyModal?: object;
-  link?: boolean;
-}
-
-function ToastComponent({
-  icon,
-  text,
-  button,
-  close,
-  type,
-  openModal,
-  openPrivacyModal,
-  link,
-}: ToastProps) {
-  const { sendEmailVerification } = useAuthentication();
-  const [toastVisible, setToastVisibility] = useState<boolean>(true);
-  const [chromeLink, setChromelink] = useState("");
-  const [fxLink, setFxlink] = useState("");
-  const [browserType] = useState(detectBrowser());
-  const { allStudies } = useStudies();
-
-  useEffect(() => {
-    allStudies.forEach((study) => {
-      if (study.authors.name === "Mozilla Rally") {
-        setChromelink(study.downloadLink.chrome);
-        setFxlink(study.downloadLink.firefox);
-      }
-    });
-  }, [chromeLink, fxLink]);
-
-  return (
-    <Alert className={type} isOpen={toastVisible}>
-      <div className="d-flex align-items-center justify-content-between">
-        <div className="left d-flex">
-          <img
-            width="24px"
-            height="24px"
-            className="warning-icon"
-            src={icon}
-            alt="warning icon"
-          />
-          <div>{text}</div>
-        </div>
-
-        <div className="right d-flex align-items-center">
-          <Button
-            onClick={async () => {
-              openModal ? await sendEmailVerification() : null;
-              openModal && (openModal as () => void)();
-              openPrivacyModal && (openPrivacyModal as () => void)();
-              setToastVisibility(false);
-            }}
-            className={`${ProductButton} toast-btn`}
-          >
-            {link ? (
-              <a
-                className={LinkStyles.NoUnderline}
-                href={browserType === BrowserType.Chrome ? chromeLink : fxLink}
-                target="_blank"
-                rel="noreferrer"
-              >
-                {button}
-              </a>
-            ) : (
-              button
-            )}
-          </Button>
-          {close && (
-            <img
-              onClick={() => setToastVisibility(false)}
-              className="close-icon"
-              src={close}
-              alt="x icon"
-            />
-          )}
-        </div>
-      </div>
-    </Alert>
   );
 }
 
@@ -155,25 +75,5 @@ const styles = {
     top: Spacing.Medium,
     width: "1226px",
     zIndex: 100,
-    $nest: {
-      ".verify-email": {
-        backgroundColor: Colors.ColorYellow100,
-      },
-      ".privacy-policy, .add-extension": {
-        backgroundColor: Colors.ColorRed100,
-      },
-
-      ".success-extension, .success-privacy": {
-        backgroundColor: Colors.ColorGreen10,
-      },
-      ".warning-icon, .toast-btn": {
-        marginRight: Spacing.xSmall,
-      },
-      ".close-icon": {
-        width: Spacing.xLarge,
-        height: Spacing.xLarge,
-        cursor: "pointer",
-      },
-    },
   }),
 };
